Reestructuracion del codigo en modulos, esqueleto basico de la clase Wfc3D
