+++ Changelog.md
@@ -1,0 +1,4 @@
+# Changelog
+
+* Reestructuracion del codigo separando en modulos cada clase
+* se agrego la estructura bacia de la clase Wfc3D para la implementacion del algoritmo WFC
