+++ src/Wfc3D.cpp
@@ -1,0 +1,43 @@
+#include <Wfc3D.hpp>
+#include <tuple>
+#include <iostream>
+
+void Wfc3D::initializeMatrix3D() noexcept
+{
+    // TODO
+}
+
+std::tuple<int, int, int> Wfc3D::findMinEntropyCell()
+{
+    // TODO
+}
+
+int Wfc3D::collapseCell(std::tuple<int, int, int> cell)
+{
+    // TODO
+}
+
+bool Wfc3D::is3DCompatible(std::tuple<int, int, int> cell1, std::tuple<int, int, int> cell2, int pattern1, int pattern2)
+{
+    // TODO
+}
+
+void Wfc3D::propagate(std::tuple<int, int, int> cell)
+{
+    // TODO
+}
+
+bool Wfc3D::executeWfcEd()
+{
+    // TODO
+}
+
+void Wfc3D::printMatrix3D() const
+{
+    // TODO
+}
+
+void Wfc3D::toGraphic() const
+{
+    // TODO
+}
