+++ src/main.cpp
@@ -1,7 +1,8 @@
 
 #include <iostream>
-// #include "Objects3D.hpp"
+// #include "Object3D.hpp"
 #include "Scene3D.hpp"
+
 int main()
 {
     std::cout << "Hello, World!" << std::endl;
